requirements.txt
tqdm
numpy
jaconv
tensorflow-gpu==1.15.4
